# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: Fantastic Coffee (decaffeinated)
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
paths:
...
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation details
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: string
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        conversationId:
                          type: string
                        senderId:
                          type: string
                        text:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        deleted:
                          type: boolean
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /messages:
    post:
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                conversationId:
                  type: string
                text:
                  type: string
                  minLength: 1
                isGroup:
                  type: boolean
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                type: object
                required: [conversationId, messageId]
                properties:
                  conversationId:
                    type: string
                  messageId:
                    type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/members:
    post:
      operationId: addToGroup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId]
              properties:
                memberId:
                  type: string
      responses:
        "204":
          description: Member added
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented yet
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented yet
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/name:
    put:
      operationId: setGroupName
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
      responses:
        "501":
          description: Not implemented yet
        "500": { $ref: "#/components/responses/InternalServerError" }
  /me/photo:
    put:
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Photo updated successfully
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413":
          description: Payload Too Large
        "415":
          description: Unsupported Media Type
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users:
    get:
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of users matching search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [name]
                  properties:
                    name:
                      type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      operationId: deleteMyPhoto
      security:
        - bearerAuth: []
      responses:
        "501":
          description: Not implemented yet
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/photo:
    put:
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Photo updated successfully
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413":
          description: Payload Too Large
        "415":
          description: Unsupported Media Type
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteGroupPhoto
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented yet
        "500": { $ref: "#/components/responses/InternalServerError" }
components:
...
  responses:
    Unauthorized:
      description: The access token is missing or it's expired
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
    Forbidden:
      description: The user is not allowed to perform this action
    NotFound:
      description: The requested resource was not found
  securitySchemes:
    LegacyAuth:
      type: apiKey
      in: header
      name: X-App-Token
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
    bearerAuth:
      type: http
      scheme: bearer
...